[REF] Rename nfe.py
