+++ l10n_br_nfe/models/__init__.py
@@ -2,7 +2,7 @@
 
 from . import res_partner
 from . import res_company
-from . import nfe
+from . import document
 from . import res_config_settings
 from . import cfop
 from . import document_cancel
